chore: corrected the project structure
