+++ package.json
@@ -24,8 +24,10 @@
     "@nestjs/core": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
     "@nestjs/typeorm": "^10.0.0",
+    "bcrypt": "^5.1.1",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
+    "jsonwebtoken": "^9.0.2",
     "mysql2": "^3.6.3",
     "owasp-password-strength-test": "^1.3.0",
     "pg": "^8.11.3",
