+++ src/app.module.ts
@@ -2,22 +2,24 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { UserEntity } from './users/entities';
+import { LoginRecordEntity, UserEntity } from './users/entities';
 import { UsersModule } from './users/users.module';
+import { AuthModule } from './auth/auth.module';
 
 @Module({
   imports: [
     TypeOrmModule.forRoot({
       type: 'postgres',
       host: 'localhost',
-      port: 3306,
-      username: 'root',
-      password: 'root',
-      database: 'test',
-      entities: [UserEntity],
+      port: 5432,
+      username: 'akhio',
+      password: 'hasonkay05',
+      database: 'university_media',
+      entities: [UserEntity, LoginRecordEntity],
       synchronize: true,
     }),
     UsersModule,
+    AuthModule,
   ],
   controllers: [AppController],
   providers: [AppService],
