+++ src/auth/auth.module.ts
@@ -1,0 +1,11 @@
+import { Module } from '@nestjs/common';
+import { AuthController } from './auth.controller';
+import { UsersService } from '../users/users.service';
+import { AuthService } from './services';
+
+@Module({
+  controllers: [AuthController],
+  providers: [AuthService],
+  imports: [UsersService],
+})
+export class AuthModule {}
