+++ src/auth/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './login.dto';
+export * from './signup.dto';
