+++ src/auth/error-handlers/index.ts
@@ -1,0 +1,1 @@
+export * from './password-is-weak.error';
