+++ src/auth/error-handlers/password-is-weak.error.ts
