--- src/users/dto/index.ts
@@ -1,2 +1,0 @@
-export * from './login.dto';
-export * from './signUp.dto';
