+++ src/users/entities/index.ts
@@ -1,2 +1,1 @@
 export * from './User.entity';
-export * from './LoginRecord.entity';
