+++ src/users/error-handlers/index.ts
@@ -1,4 +1,3 @@
 export * from './email-is-taken.error';
 export * from './password-doesnt-match.error';
-export * from './password-is-weak.error';
 export * from './user-not-found.error';
