--- src/users/interfaces/index.ts
@@ -1,1 +1,0 @@
-export * from './authenticated-user.interface';
