--- src/users/services/index.ts
@@ -1,3 +1,0 @@
-export * from './auth.service';
-export * from './login-audit.service';
-export * from './users.service';
