+++ src/users/users.module.ts
@@ -1,12 +1,12 @@
 import { Module } from '@nestjs/common';
-import { UsersController } from './users.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { UserEntity } from './entities';
-import { UsersService } from './services/users.service';
+import { UsersService } from './users.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([UserEntity])],
-  controllers: [UsersController],
+  controllers: [],
   providers: [UsersService],
+  exports: [UsersService],
 })
 export class UsersModule {}
